Select the transformation type based on the number of windows
The type of transform was fixed as affine, leading to nonsense results
if fewer than three focus windows were selected. The transform is now
chosen to exactly match the DOF specified by the number of windows if
four or fewer points are chosen, or using projective transformation 
and the least square fit from skimage's estimate transform for more
than four points.

+++ aperturesynth/synthesise.py
@@ -47,7 +47,7 @@ def save_image(image, filename):
 
 def load_image(image):
     """Loads the given file and converts to float32 format. """
-    return img_as_float(io.imread(image)).astype('float32')
+    return img_as_float(io.imread(image))
 
 
 def register_images(image_list, registrator):
